+++ app/[lang]/(app)/(home)/(tablet)/room-devision/rooms-rack/components/rooms/RackRooms.tsx
@@ -29,12 +29,13 @@ export default function RackRooms({
    </div>
   );
 
+ console.log(rack.data);
  return (
   <div className='grow '>
    {rack.isLoading && <LinearLoading />}
    <div
     data-layout-minimal={rack.rackView === 'minimal'}
-    className='grid gap-4 grid-cols-[repeat(auto-fill,minmax(9rem,9.5rem))] data-[layout-minimal=true]:grid-cols-[repeat(auto-fill,minmax(6.5rem,7rem))] justify-center pb-4 data-[layout-minimal=true]:gap-3'
+    className='grid gap-3 grid-cols-[repeat(auto-fill,minmax(9rem,9.5rem))] data-[layout-minimal=true]:grid-cols-[repeat(auto-fill,minmax(5.8rem,6rem))] justify-center pb-4 data-[layout-minimal=true]:gap-2'
    >
     {rack.data.map((room) => (
      <RackRoom