+++ app/[lang]/(app)/(home)/(tablet)/room-devision/rooms-rack/utils/room-notes/noteTypes.ts
@@ -1,0 +1,7 @@
+export enum NoteTypes {
+ normal = 1,
+ important,
+ veryImportant,
+ urgent,
+ onCheckout,
+}