+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - add res salons user accessibility (hra-done)
 - add apptype request header field (hra-done)
 - add check the right path (hra-done)
+- add rack notif center
 
 # 2.0.5
 ### features